♻️ refactor: 경기등록 type 정리 #63
