+++ src/pages/admin/type.ts
@@ -2,7 +2,7 @@ export interface FormValueType {
   stadiumName: string;
   date: string;
   sport: string;
-  homeTeam: string;
-  awayTeam: string;
+  homeTeamName: string;
+  awayTeamName: string;
   gameStartTime: string;
 }
